% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP_Dataframe.R
\name{aggregateByLabel}
\alias{aggregateByLabel}
\title{Label a list of int attractors and agregate by label}
\usage{
aggregateByLabel(df, node.names, label.rules, sep = "/")
}
\arguments{
\item{df}{dataframe with states as rownames}

\item{node.names}{node names of the state, the length must be the same that the state's}

\item{label.rules}{dataframe with labels (1st col) and rules (2nd col), if more than one rule is true all labels are appendedl the node names present in the rules must be in node.names}

\item{sep}{string to join elements with length > 1, default "/"}
}
\value{
Dataframe, each row corresponds to a label and each column corresponds to a property of the original dataframe
}
\description{
Takes a dataframe where the rownames are states in integer format (cycles are strings joined with sep). The function labels the states and agregates the dataframe using label. Agregate splits the data into subsets by label, computes summary statistics for each, and returns the result in a convenient form. See   \code{\link[stats]{aggregate}}
}
\keyword{internal}
