% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP_Dataframe.R
\name{attractorsToDataframe}
\alias{attractorsToDataframe}
\title{Convert a list of attractors to a data frame.}
\usage{
attractorsToDataframe(attr.list, sep = "/",
  returnDataFrame = c("occurrence", "basinSize"), ...)
}
\arguments{
\item{attr.list}{list of BoolNet attractor objects}

\item{sep}{string to join elements with length > 1, default "/"}

\item{returnDataFrame}{if returnDataFrame='occurrence' returns a df where each column corresponds to a network and the rows to the attractor/label or labels. The values indicate the frequency of the attractor/label
if returnDataFrame='basinSize' returns a df where the values indicate the basin size of the attractor/label
if returnDataFrame='attrList' returns a list of AttractorInfo objects}
}
\value{
Dataframe, each column corresponds to a property of the attractor
}
\description{
Convert a list of BoolNet attractor objects to a data frame. Each property of each attr$attractors corresponds to a dataframe column. Columns are named attrName.propertyName, if the list has no names numbers will be used. If the property has elements with length > 1 it converts them to a string and joins them with sep.
}
\examples{
data(cellcycle)
attrs <- list(getAttractors(cellcycle))
attractorsToDataframe(attrs)

}
\keyword{internal}
