% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP_Helper.R
\name{bin2intState}
\alias{bin2intState}
\title{Convert binart state vector to integer.
'}
\usage{
bin2intState(x)
}
\arguments{
\item{x}{input vector representing the state}
}
\value{
Corresponding int.
}
\description{
Convert binart state vector to integer.
'
}
\examples{
bin2intState( c(0,1,0,0,0,1,0,1,0,0) )
}
\seealso{
\code{\link{intToBits}} which this function wraps
}
