% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP_Dataframe.R
\name{countChangesDataframe}
\alias{countChangesDataframe}
\title{Count won or lost values in comparison with those present in a reference column.}
\usage{
countChangesDataframe(df, reference = "WT", axis = 2)
}
\arguments{
\item{df}{dataframe with numeric values}

\item{reference}{name of column to use as reference}
}
\value{
axis, axis to obtain new/lost values, see \code{\link[base]{sum}}
}
\description{
Counts how many values were won or lost in comparison between dataframe. It measures ocurrences and then counts the values with sum.
}
\examples{
df <- data.frame(WT=c(1,2,0,0), a=c(1,2,0,0), 
                 b=c(0,2,1,0), c=c(1,0,3,1))
countChangesDataframe(df)
#     	WT	    a	b	c
#     	new	    0	0	1	2
#     	lost	0	0	1	1
countChangesDataframe(df, axis=1)
#     	WT	    a	b	c
#     	new	    0	0	2	1
#     	lost	1	1	0	0

}
\keyword{internal}
