% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolNetPerturb.R
\name{dec2binState}
\alias{dec2binState}
\title{Convert integer to binary state vector with node names.
'}
\usage{
dec2binState(x, nodes)
}
\arguments{
\item{x}{input integer representing the state}

\item{genes}{network node names}
}
\value{
Numeric binary vector of the same length as the nuber of nodes. Each 
    position corresponds to a node of the network. The values of each element 
    are 0 or 1. The name of each element corresponds to the name of the node in
    that position in the network.
}
\description{
Convert integer to binary state vector with node names.
'
}
\examples{
dec2binState(0, net$genes)
}
\seealso{
\code{\link{intToBits}} which this function wraps
}

