% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP_PerDerrida.R
\name{derrida}
\alias{derrida}
\title{Derrida analysis.}
\usage{
derrida(net, repetitions = 1000)
}
\arguments{
\item{net}{BoolNet network to perturb}

\item{repetitions}{number of repetitions}
}
\value{
vector with average hamming distance in t+1, names are hamming distance in t
}
\description{
Derrida curves are used to determine the dynamic behaviour of Boolean networks. We take two states with Hamming distance h in the time t, and determine the hamming distance in time t+1. This process is repeated multiple times. If the network is chaotic, the curve tends to be over the diagonal. If the network is ordered it will be below the diagonal.
}
\examples{
data(cellcycle)
derrida(cellcycle)

}
