% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP-EpigeneticLandscape.R
\name{epigeneticLandscape}
\alias{epigeneticLandscape}
\title{Calculate the epigenetic landscape of a network}
\usage{
epigeneticLandscape(net, p = 0.01, returnDataFrame = F)
}
\arguments{
\item{net}{BoolNet network to perturb}

\item{p}{probability of noise in the simulation}

\item{returnDataFrame}{returns a df where ...}
}
\value{
The transition probabilities between states.
}
\description{
...
}
\examples{
data(netTh17Treg)
epigeneticLandscape(netTh17Treg, p = 0.01, returnDataFrame = T)
}
