% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP_Label.R
\name{labelState}
\alias{labelState}
\title{Label a state.}
\usage{
labelState(state, node.names, label.rules, sep = "")
}
\arguments{
\item{state}{binary state to label}

\item{node.names}{node names of the state, the length must be the same that the state's}

\item{label.rules}{dataframe with labels (1st col) and rules (2nd col), if more than one rule is true all labels are appendedl the node names present in the rules must be in node.names}

\item{sep}{string to separate the labels when more than one can be applied to the state.}
}
\value{
String corresponding to the label of the state.
}
\description{
Labels a binary state using a set of labelling rules.
The elements of the state correspond the network node names. If a rule is satisfied the corresponding label is appended.
}
\examples{
labelState(state, net$genes, label.rules, sep='')
}
