% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP_PerFunctions.R
\name{perturbNetworkFixedNodes}
\alias{perturbNetworkFixedNodes}
\title{Perturb network functions knock-out (0) or over-expression (1).}
\usage{
perturbNetworkFixedNodes(net, genes, values = 0, names,
  returnDataFrame = c("occurrence", "basinSize", "attrList"),
  label.rules = NULL, sep = "/", verbose = FALSE, ...)
}
\arguments{
\item{net}{network to perturb}

\item{genes}{list of gene names to perturb. To perturb multiple nodes at the same time use a vector inside the list.}

\item{values}{list of values of the perturbed genes. Knock-out is 0, over-expression is 1.}

\item{names}{names of the perturbation}

\item{returnDataFrame}{if returnDataFrame='occurrence' returns a df where each column corresponds to a network and the rows to the attractor/label or labels. The values indicate the frequency of the attractor/label
if returnDataFrame='basinSize' returns a df where the values indicate the basin size of the attractor/label
if returnDataFrame='attrList' returns a list of AttractorInfo objects}

\item{label.rules}{if label.rules=NULL (default) returns a df with the attractors as rows
if label.rules is a dataframe with labels (1st col) and rules (2nd col), if more than one rule is true all labels are appendedl the node names present in the rules must be in node.names}

\item{sep}{string to join states in cyclic attractors, default "/"}

\item{verbose}{report the node being perturbed, useful for larger networks}

\item{...}{Further parameters to getAttractors.}
}
\value{
dataframe or list of attractors of the perturbed networks
}
\description{
Simulates fixed function perturbations (knock-out and over-expression) in a network. Takes a network, a list of gene names and corresponding values to perturb, fixes those values in each network and returns the corresponding attractors for all perturbed networks. By default it  returns all the single node knock-out and over-expression of a network.
}
\examples{
data(cellcycle)
# All single gene knock-out and over-expression of cellcycle network
perturbNetworkFixedNodes(cellcycle)

# Cellcycle network CycD over-expression
perturbNetworkFixedNodes( cellcycle, list("CycD"), list(1) )

# Cellcycle network CycD knock-out with Rb over-expression
perturbNetworkFixedNodes(  cellcycle, list(c("CycD", "Rb")), list(c(0,1)) , list("CycD-RB+") )

# All CycD double gene knock-outs
double.KO <- lapply(cellcycle$genes[-1], function(gene) c("CycD", gene))
number.double.KO <- length(double.KO)
perturbNetworkFixedNodes(  net=cellcycle, genes=double.KO, 
                        value=as.list(rep(0, number.double.KO))  ) 
}
\seealso{
\code{\link{fixGenes}} \code{\link{attractorListToDataframe}}
}
