% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP_Population.R
\name{populationSimulation}
\alias{populationSimulation}
\title{Analyze and graph the average of multiple asynchronous simulations for an specific number of time steps}
\usage{
populationSimulation(net, tStates = 1000, simulations = 1000,
  new.nodes = list(), new.values = list(), gene = NULL,
  duration = 1, value = 1, pertTime = 400, onePert = TRUE)
}
\arguments{
\item{net}{Boolnet network}

\item{tStates}{Number of time steps running each simulation, default value set to 1000.}

\item{simulations}{Number of simulations equivalent to "population size", default value set to 1000.}

\item{new.nodes}{Nodes to fix with an specific value at the initial state.}

\item{new.values}{Values for nodes in new.nodes.}

\item{gene}{Gene to perturb at an specific time-step (pertTime), default value set NULL.}

\item{duration}{Time of perturbation, default value set to 1.}

\item{value}{Value of perturbed gene, default value set to 1.}

\item{pertTime}{Specific time-step at which the gene will be perturbed, default value set to 400.}

\item{onePert}{If this is TRUE the perturbation will only occur one time during the simulation. If FALSE, de perturbation wil occur periodically, default value set to TRUE.}
}
\value{
Data frame with the average of the simulations for each node in each time step
}
\description{
Analyze and graph the average of multiple asynchronous simulations for an specific number of time steps
}
\examples{
data(netTh17Treg)
pop_Th17 <- popSimulation(netTh17Treg, new.nodes = c("IL2e","IL21e","TGFBe"), new.values = c(1,1,1), gene ="TGFBe", value = 0)

}
