% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP_Summarize.R
\name{summarizeAttractors}
\alias{summarizeAttractors}
\title{Summarize attractors of a dataframe ignoring certain nodes.}
\usage{
summarizeAttractors(attr, genes, returnBasin = F)
}
\arguments{
\item{attr}{BoolNet attractor object}

\item{genes}{list of gene names to ignore, by default it detects inputs. To perturb multiple nodes at the same time use a vector inside the list.}

\item{returnBasin}{default FALSE, return a column with the combined basin size}
}
\value{
dataframe, each column corresponds to the number of attractor, state, or node. If Boolean=FALSE return dataframe, each column corresponds to a property of the attractor.
}
\description{
...
}
\examples{
data(netTh17Treg)
attr <- getAttractors(netTh17Treg)
summarizeAttractors(attr)

}
