% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP_Helper.R
\name{validateState}
\alias{validateState}
\title{Validate and convert a number to a valid state.}
\usage{
validateState(x, node.names)
}
\arguments{
\item{x}{input integer representing the state, can be a vector of 0/1 or an int}

\item{node.names}{network node names}
}
\value{
Numeric binary vector of the same length as the number of nodes. Each 
    position corresponds to a node of the network. The values of each element 
    are 0 or 1. The name of each element corresponds to the name of the node in
    that position in the network.
}
\description{
Validate and convert a number to a valid state.
}
\examples{
validateState(162, cellcycle$genes)
validateState('162', cellcycle$genes)
validateState(c(0,1,0,0,0,1,0,1,0,0), cellcycle$genes)
}
\seealso{
\code{\link{intToBits}} which this function wraps
}
