% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP_Perturbations.R
\name{verifySyncronousVsAsyncronous}
\alias{verifySyncronousVsAsyncronous}
\title{Compare atractors with synchronous and asynchronous updating.}
\usage{
verifySyncronousVsAsyncronous(net, attr, label.rules = NULL, ...)
}
\arguments{
\item{net}{BoolNet network to perturb}

\item{attr}{BoolNet AttractorInfo object}

\item{label.rules}{dataframe with labels (1st col) and rules (2nd col), if more than one rule is true all labels are appendedl the node names present in the rules must be in node.names}

\item{...}{Further parameters to  \code{\link{getAttractors}}.}
}
\value{
dataframe comparing synchronous and asynchronous updating
}
\description{
Takes a set of BoolNet attractor or states and verifies if they can be recovered with asynchronous updating. See \code{\link[BoolNet]{getAttractors}} with options type='async' and method="chosen".
}
\examples{
data(cellcycle)
attr <- getAttractors(cellcycle)
verifyAttractorSyncronousVsAsyncronous(cellcycle, attr)
}
\seealso{
\code{\link[BoolNet]{getAttractors}}
}
