% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolNetPerturb.R
\name{perturbNetworkFixedNodes}
\alias{perturbNetworkFixedNodes}
\title{Perturb network functions.}
\usage{
perturbNetworkFixedNodes(net, genes, value, label, type = "synchronous",
  returnDataFrame = TRUE)
}
\arguments{
\item{net}{network to perturb}

\item{genes}{list of gene names to perturb. To perturb multiple nodes at the same time use a vector inside the list.}

\item{value}{list of values of the perturbed genes. Knock-out is 0, over-expression is 1.}

\item{label}{name of the perturbation}

\item{type}{update type, can be "synchronous" (default) or "synchronous"}

\item{returnDataFrame}{if TRUE returns a dataframe where the rownames correspond to the states and columns correspond to the attractor basin size of the perturbed network, if FALSE returns a list of BoolNet attractors}
}
\value{
dataframe or list of attractors of the perturbed networks
}
\description{
Simulates fixed function perturbations (knock-out and over-expression) in a network. Takes a network, a list of gene names and corresponding values to perturb, fixes those values in each network and returns the corresponding attractors for all perturbed networks. By default it  returns all the single node knock-out and over-expression of a network.
}
\examples{
# All single gene knock-out and over-expression of a network
perturbNetworkFixedNodes(net)

# Cellcycle network CycD over-expression
perturbNetworkFixedNodes( cellcycle, list("CycD"), list(1) )

# Cellcycle network CycDknock-out with Rb over-expression
perturbNetworkFixedNodes(  cellcycle, list(c("CycD", "Rb")), list(c(0,1)) , list("CycD-RB+") )

# All CycD double gene knock-outs
double.KO <- lapply(cellcycle$genes[-1], function(gene) c("CycD", gene))
number.double.KO <- length(double.KO)
perturbNetworkFixedNodes(  net=cellcycle, genes=double.KO, 
                        value=as.list(rep(0, number.double.KO))  ) 
}
\seealso{
\code{\link{fixGenes}}
}

